# Annotate Kernel#class as inline