--- a/benchmark/mjit_send_cfunc.yml
+++ b/benchmark/mjit_send_cfunc.yml
@@ -1,7 +1,0 @@
-type: lib/benchmark_driver/runner/mjit
-prelude: |
-  def mjit_send_cfunc
-    self.class
-  end
-benchmark: mjit_send_cfunc
-loop_count: 100000000